chore: translate error message
